safari-534.53-branch/LayoutTests/sputnik/Conformance/08_Types/8.3_The_Boolean_Type/S8.3_A2.2-expected.txt

S8.3_A2.2

PASS ReferenceError: Left side of assignment is not a reference.

TEST COMPLETE
